% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{add_date_cols}
\alias{add_date_cols}
\title{Add date information}
\usage{
add_date_cols(df, dt_col, cols)
}
\arguments{
\item{df}{data frame that has the date column}

\item{dt_col}{name of the column with the dates of interest}

\item{cols}{columns to add, currently "year","yr","month","week","day", "weekday","month_name","month_abb","weekday_name","weekday_abb"}
}
\value{
same data frame with added columns
}
\description{
Add columns to a data frame with information extracted from the indicated date column
}
\examples{

df<-data.frame(date=sample(seq(as.Date('2010/01/01'), as.Date('2019/12/31'), by="day"), 30))
add_date_cols(df,"date")

}
