% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{as.excel.dates}
\alias{as.excel.dates}
\title{Convert excel integer dates to R dates}
\usage{
as.excel.dates(dt_col)
}
\arguments{
\item{dt_col}{}
}
\description{
Convert excel integer dates to R dates
}
