% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_like.R
\name{as_excel_dates}
\alias{as_excel_dates}
\title{Convert column of integers to excel dates}
\usage{
as_excel_dates(df, ...)
}
\arguments{
\item{df}{- data.frame}

\item{...}{- column to convert}
}
\description{
Convert column of integers to excel dates
}
