% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{chi_square}
\alias{chi_square}
\title{Pearson's Chi-squared test}
\usage{
chi_square(vals, pvalue = T, correct = FALSE)
}
\arguments{
\item{vals}{an integer vector of numerator (odd elements), denominator (even elements) pairs}

\item{pvalue}{a logical TRUE: return only p-value | FALSE: return list from chisq.test}

\item{correct}{a logical indicating whether to apply continuity correction when computing the test statistic for 2 by 2 tables}
}
\value{
numeric p-value or list with class "htest"
}
\description{
Convenience function for stats::chisq.test(). Returns only the p-value if pvalue is set to TRUE (default),
otherwise returns a list with class "htest"
}
\examples{
v<-c(c(100,144),c(200,344),c(75,130))
orrr::chi_square(v)

}

