% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dir.R
\name{convert.dot}
\alias{convert.dot}
\title{Normalize path}
\usage{
convert.dot(path)
}
\arguments{
\item{names}{character vector of path name}
}
\value{
character - path, or empty string if not successful
}
\description{
Returns a vector substituting the project folder for .
}
\examples{
\dontrun{
convert.dot("./data")
}

}
