% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{convert.excel.dates}
\alias{convert.excel.dates}
\title{Convert Excel Date to R Dates}
\usage{
convert.excel.dates(dts)
}
\arguments{
\item{dts}{- a vector of Excel dates (integers)}
}
\value{
those integer dates converted to R dates
}
\description{
Applies the correct origin (1899-12-30) to integer Excel dates
}
\examples{
\dontrun{
df<-xlsx::read.xlsx(file=fname)
df$DOB<-convert.excel.dates(df$DOB)

}
}
