% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{current_age}
\alias{current_age}
\title{Current age in integral years between two Dates}
\usage{
current_age(dob)
}
\arguments{
\item{dob}{date of birth}
}
\value{
integer - age in years
}
\description{
Returns the current age (integer) based on a date of birth. The fraction is truncated.
}
\examples{

dob<-as.Date("2000-01-01")
current_age(dob)
}
