% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dir.R
\name{dir.project.create}
\alias{dir.project.create}
\title{Create Project Directory}
\usage{
dir.project.create(names)
}
\arguments{
\item{names}{character vector of folder names in a branch}
}
\value{
character - path, or empty string if not successful
}
\description{
Returns whether or not the folder was created
}
\examples{
dir.project.create("data")

}
