% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dir.R
\name{dir.project}
\alias{dir.project}
\title{Project Directory}
\usage{
dir.project(names, slash = T)
}
\arguments{
\item{names}{character vector of folder names in a branch. if missing will return project root folder}

\item{slash}{logical append slash at end of path. Default is TRUE.}
}
\value{
character - path of the project directory/folder
}
\description{
Returns the name of the project directory by stepping out from the working directory until a *.Rproj file is found
}
\examples{
dir.project()

dir.project("data")

# following are equivalent
dir.project("data/excel/csv")
dir.project(c("data","excel","csv"))


}
