% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flextable.R
\name{ft_to_xlsx}
\alias{ft_to_xlsx}
\title{Convert a flextable object to an xlsx file}
\usage{
ft_to_xlsx(
  ft,
  file = "./output/ft_to_xl.xlsx",
  sheet_name = "ft1",
  start_row = 1,
  start_col = 1,
  append = FALSE
)
}
\arguments{
\item{append}{}
}
\description{
Convert a flextable object to an xlsx file
}
