% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rdata.R
\name{get.rdata}
\alias{get.rdata}
\title{Get Data file item}
\usage{
get.rdata(file, item = 1)
}
\arguments{
\item{file}{- character name of file}

\item{item}{- integer item to retrieve (default=1) or character matching one item (using grep)}
}
\value{
requested item in a .RData file
}
\description{
Returns the nth item in a data file
}
\examples{
dir.project()

}
