% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{getLines}
\alias{getLines}
\title{returns the lines of a file as a character vector}
\usage{
getLines(filename)
}
\arguments{
\item{filename}{- character - name of file to readLoines from}
}
\value{
character - the lines of the file minus blank lines
}
\description{
returns the lines of a file as a character vector
}
