% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_like.R
\name{is.integer_like}
\alias{is.integer_like}
\title{Can Vector be Coerced to Integer}
\usage{
is.integer_like(x)
}
\arguments{
\item{x}{integer, numeric, or character vector}
}
\value{
logical - True if is.integer or can be coerced to an integer
}
\description{
Can Vector be Coerced to Integer
}
