% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_like.R
\name{is.numeric_like}
\alias{is.numeric_like}
\title{Can Vector be Coerced to Numeric}
\usage{
is.numeric_like(x)
}
\arguments{
\item{x}{numeric or character vector}
}
\value{
logical - True if is.numeric or can be coerced to a numeric
}
\description{
Can Vector be Coerced to Numeric
}
