% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_list.R
\name{list.functions}
\alias{list.functions}
\title{List of Functions}
\usage{
list.functions(files, saveto = NULL, verbose = TRUE, sorted = TRUE)
}
\arguments{
\item{files}{character - files or folders or a mix; if blank all files and folders in ./code and ./R}

\item{saveto}{character - name of file to save (sink) the text output to}

\item{verbose}{boolean - if verbose is true, print the structure ... folder/file/function}
}
\value{
invisible(data.frame) -  with columns for folder, file, and function
}
\description{
Gets a list of the functions in a file, a folder, a vector of files or folders, or all
files in ./code or ./R (if files parameter is missing)
}
