% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_rdata.R
\name{list.rdata}
\alias{list.rdata}
\title{Data file items}
\usage{
list.rdata(file)
}
\arguments{
\item{file}{- character name of file}
}
\value{
character - vector of names of items in a .RData file
}
\description{
Returns the names of items in a data file
}
\examples{
dir.project()

}
