% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parens.R
\name{matching_parens}
\alias{matching_parens}
\title{Matching Parentheses}
\usage{
matching_parens(text)
}
\arguments{
\item{text}{}
}
\value{
data frame with matching pair positions
}
\description{
Finds matching parentheses in a string. Returns a data frame with each row representing a matching pair
of parentheses, `op` being the opening position and `cl` being the closing position. If processed in row
order, you should never see an outer set before its inner set(s)
}
\examples{
txt <- "((x + 5) * (y - 7)) / (z^2 + 4)"
matching_parens(txt)

}
