% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_parse.R
\name{parse_name}
\alias{parse_name}
\title{Retieve additional name data from the character vector containing the full name}
\usage{
parse_name(name, hyphen = F)
}
\arguments{
\item{name}{- character: the full name of the person in Meditech name format (LASTNAME [SUFFIX],FIRSTNAME MIDDLENAME [SUFFIX])}

\item{hyphen}{- logical: if TRUE, substitute hyphens for spaces in last name. Some applications prefer no spaces in last names.}
}
\value{
list : elements are: name, last_name, first_name, middle_name, Suffix
}
\description{
Retieve additional name data from the character vector containing the full name
}
\examples{
parse_name("JOHNSON SR, JAMES R.")

}
