% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_parse.R
\name{parse_suffix}
\alias{parse_suffix}
\title{Parse and Split Suffix from Last Name}
\usage{
parse_suffix(df, ln_col, sfx_col = "suffix")
}
\arguments{
\item{df}{data.frame: containing a last name column}

\item{ln_col}{character: column name containing last name}

\item{sfx_col}{character: column name to contain suffix}
}
\description{
Parse and Split Suffix from Last Name
}
\examples{

df <- data.frame(ln = c("JONES","WILLIAM III","RAFERTY SMITH","O CONNELL","HANSEN JR"),
                 fn = c("TOM","CLARENCE","RAFFIE","DONAL","CHUCKY"))
df \%>\% parse_suffix(ln_col = "ln")


}
