% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_names.R
\name{random_first_name}
\alias{random_first_name}
\title{Load Data Frame with Random First Names}
\usage{
random_first_name(df, name_col, gender_col = NULL, genders = c("F", "M"))
}
\arguments{
\item{name_col}{- column name with first name data}

\item{gender_col-}{column name with gender data}

\item{genders-}{gender text for female,male (default=c("F","M"))}
}
\description{
Load Data Frame with Random First Names
}
