% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_names.R
\name{random_last_name}
\alias{random_last_name}
\title{Load Data Frame with Random Last Names}
\usage{
random_last_name(df = NULL, name_col = NULL, n = 1)
}
\arguments{
\item{df}{data.frame - data.frame to store the last name column}

\item{name_col}{character - column name with last name data}

\item{n}{integer number of names if df is NULL}
}
\description{
Load Data Frame with Random Last Names
}
