% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{read.access.table}
\alias{read.access.table}
\title{Read Access Table}
\usage{
read.access.table(file, tbl, pwd = "")
}
\arguments{
\item{file}{character path/file name of MS Access database file (.accbd, .mdb)}

\item{tbl}{character Name of the table to fetch - if missing, returns the name of all tables}

\item{pwd}{character Password, default is ""}
}
\value{
character - path of the project directory/folder
}
\description{
Returns a data.frame with the contents of a MS Access table
}
\examples{
\dontrun{
file<-"C:/Data/xyz.accdb"
tbl<-"tblSomeData"
df<-read.access.table(file,tbl,pwd="uJ80s$R")


}
}
