% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{search_r_files}
\alias{search_r_files}
\title{R File Search}
\usage{
search_r_files(
  find = NULL,
  ext = "R",
  r_root = NULL,
  file_color = "#4444ff",
  results_color = "red",
  date_color = "#669977",
  results_bg = "#f7f7f7",
  hyph_count = 90,
  ignore_comments = TRUE,
  ...
)
}
\arguments{
\item{find}{- character: text to search for}

\item{ext}{- character - extension (default: .R)}

\item{r_root}{- character - folder root for searching for .R files}

\item{results_bg}{- character - results bg color (use #hhhhhh format )}

\item{ignore_comments}{- logical - ignore all lines that are comments (start with #)}

\item{...}{- unused}

\item{file_col}{- character - filename text color (use #hhhhhh format )}

\item{results_col}{- character - results text color (use #hhhhhh format )}

\item{date_col}{- character - file date text color (use #hhhhhh format )}
}
\value{
vector of results
}
\description{
R File Search
}
\examples{
search_r_files("foobar")


R File Search

 Sends formatted text to command line showing results for the search

search_r_files("foobar")
}
