% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{source_lib}
\alias{source_lib}
\title{sources the file (usually a script file of r functions)}
\usage{
source_lib(
  libname,
  proj_path = orrr::dir.project(),
  lib_path = "code/libs/",
  prefix = "lib_",
  verbose = FALSE
)
}
\arguments{
\item{libname}{- character - name of file to readLoines from}

\item{proj_path}{- character - path to the project,  passed to orrr::dir.project; default value= "orrr::dir.project()"}

\item{lib_path}{- character - path to the project,  passed to orrr::dir.project; default value= "code/libs/"}

\item{prefix}{- character - prefix added to lib to get filename; default value = "lib_"}

\item{verbose}{- logical - cat the filename; default value = FALSE}
}
\value{
the value of the sourcing of the file
}
\description{
sources the file (usually a script file of r functions)
}
