% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_utils.R
\name{string_ruler}
\alias{string_ruler}
\title{String ruler
Fuction to assist with visualizing positions of a string, usually used with long fixed width
character vectors}
\usage{
string_ruler(x, split = 0, under = T, uchar = "-", nl = 0,
  nlchar = " ")
}
\arguments{
\item{x}{- character: string to measure}

\item{split}{- integer: split lines every [split] characters}

\item{under}{- logical: underline the ruler digits}

\item{uchar}{- character: character to use for the underline (default="-")}

\item{nl}{- integer: number of newlines after each ruled line}

\item{nlchar}{- character: character to use for the new lines (default=" ")}
}
\value{
character: string with "ruler" above
}
\description{
String ruler
Fuction to assist with visualizing positions of a string, usually used with long fixed width
character vectors
}
