% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\docType{data}
\name{x}
\alias{x}
\title{R File Search}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
x
}
\arguments{
\item{find}{- character: text to search for}
}
\value{
vector of results
}
\description{
R File Search
}
\examples{
search_r_files("foobar")


}
\keyword{datasets}
