% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{years_between}
\alias{years_between}
\title{Number of integral years between two Dates}
\usage{
years_between(date1, date2)
}
\arguments{
\item{date1}{date earlier date}

\item{date2}{date later date}
}
\value{
integer - number of years between date1 and date2
}
\description{
Returns the number of years (integer) between two dates. The fraction is truncated. (like age)
}
\examples{

dt1<-as.Date("2000-01-01")
years_between(dt1,Sys.Date())
}
