% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{add_age_intervals}
\alias{add_age_intervals}
\title{Add Age Intervals}
\usage{
add_age_intervals(
  df,
  breaks = NULL,
  age_col = "Age",
  group_col = "AgeGroup",
  under = "Younger than",
  over = "or Older"
)
}
\arguments{
\item{df}{- data.frame}

\item{breaks}{- integer - vector of age intervals (passed to cut)}

\item{age_col}{- character - name of column with numeric ages}

\item{group_col}{- character - name of factor column with text age ranges}

\item{under}{- character - text to use for less/younger than}

\item{over}{- character - text to use for greater/older than}
}
\value{
data.frame
}
\description{
Add a column (factor) with plain language description of the age range
}
