% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_utils.R
\name{center_string}
\alias{center_string}
\title{Center a string in a string with a given length by padding the ends}
\usage{
center_string(x, len_out, ctr_char)
}
\arguments{
\item{ctr_char}{}
}
\description{
Center a string in a string with a given length by padding the ends
}
