% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_utils.R
\name{colnames.pretty}
\alias{colnames.pretty}
\title{Pretty Column Names
this prints a vertical list of column names
  that are quoted and separated by commas
   useful for getting a usable list for copying and pasting}
\usage{
colnames.pretty(df, quotes = TRUE, vert = F, sep = ",")
}
\arguments{
\item{df}{- a data.frame}

\item{quotes}{- logical - names are quoted?}

\item{vert}{- logical - separate names by a newline?}

\item{sep}{- character - character to separate names by, default = ","}
}
\value{
- nothing
}
\description{
Pretty Column Names
this prints a vertical list of column names
  that are quoted and separated by commas
   useful for getting a usable list for copying and pasting
}
\examples{
colnames.pretty(mtcars)
colnames.pretty(mtcars,vert=TRUE)

}
