% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dir.R
\name{dir.project}
\alias{dir.project}
\title{Project Directory}
\usage{
dir.project(names, slash = T)
}
\arguments{
\item{names}{character vector of folder names in a branch. If missing, this function
will return project root folder}

\item{slash}{logical append slash at end of path. Default is TRUE.}
}
\value{
character - path of the project directory/folder
}
\description{
Returns the path of the project directory and adds and other folders passed
 as the names argument
}
\examples{
\dontrun{
dir.project()

dir.project("data")

# following are equivalent
dir.project("data/excel/csv")
dir.project(c("data","excel","csv"))
}

}
