% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_dir.R
\name{dir.project}
\alias{dir.project}
\title{Project Directory}
\usage{
dir.project(slash = F)
}
\value{
character - name of trhe project directory/folder
}
\description{
Returns the name of the project directory by stepping out from the working directory until a *.Rproj file is found
}
\examples{
dir.project()

}

