% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{search_r_files}
\alias{search_r_files}
\title{R File Search}
\usage{
search_r_files(find = NULL)
}
\arguments{
\item{find}{- character: text to search for}
}
\value{
vector of results
}
\description{
R File Search
}
\examples{
search_r_files("foobar")

}
